move PaymentStatus into route as well
