+++ api/src/routes/payment/index.ts
@@ -4,6 +4,11 @@ import './finalize.js'
 import './status.js'
 
 export type { PaymentQuoteInput } from './quote.js'
-export type { PaymentStatusParam } from './status.js'
+export type {
+  PaymentStatusParam,
+  PaymentStatus,
+  PaymentStatusRejected,
+  PaymentStatusSuccess,
+} from './status.js'
 export type { PaymentGrantInput } from './grant.js'
 export type { PaymentFinalizeInput } from './finalize.js'
