+++ api/src/types.ts
@@ -3,6 +3,9 @@ export type {
   PaymentGrantInput,
   PaymentQuoteInput,
   PaymentStatusParam,
+  PaymentStatus,
+  PaymentStatusRejected,
+  PaymentStatusSuccess,
 } from './routes/payment'
 
 export type ApiErrorResponse = {
@@ -15,16 +18,3 @@ export type ApiErrorResponse = {
     }
   }
 }
-
-export type PaymentStatusSuccess = {
-  paymentId: string
-  hash: string
-  interact_ref: string
-}
-
-export type PaymentStatusRejected = {
-  paymentId: string
-  result: 'grant_rejected'
-}
-
-export type PaymentStatus = PaymentStatusSuccess | PaymentStatusRejected
